Initial implementation of dynamc day/date meshes.
Each part of the Day/Date (that is the day, the month, and the date) are
now generated independently and automatically placed relative to each
other and the end of the day/date box.

This will allow more control over the text spacing, and could allow for
eventual multi-language support.